requirements.txt
azure-servicebus
faker
